JoystickConfigController: support 2 simultaneous joystick configuration:
Prepare this class for the upcomming multi-joystick support.

As this class is instantiated from QML depending on the user entering
the vehicle setup->joystick submenu, we need to tell it from there if
we are referring to the first or second joystick configuration.

For this matter, some initialization was moved from constructor to the
new init() function, and that init function must be called from qml after
The object has been instantiated, and after optionally setting the
issecondary property ( false by default ).