Joystick.cc: on _setDefaultCalibration, don't set _calibrated = true:
This is meant to provide automatic calibration, axis mapping, etc
for SDL Joysticks detected as game controllers. Most of the time
this is good, and saves the user the calibration time, but under
some circunstances the mapping might not be correct, and it can
provoke a dangerous situation where the user thinks Joystick is
alright when it is not.