FlightDisplay: remove gimbal yaw lock call
We don't need this separate yaw lock call as the flags are already sent
as part of sendGimbalManagerPitchYaw which uses follow mode.

This caused problems because it leads to the situation where two
consequtive GIMBAL_MANAGER_DO_PITCH_YAW commands are sent, first for the
flags, second for the pitch and yaw values. The second one is dropped
with a warning because the "same command" is still pending.

Signed-off-by: Julian Oes <julian@oes.ch>